+++ js/initial-data.js
@@ -1,0 +1,8 @@
+const checkins = ['12:00', '13:00', '14:00'];
+const checkouts = ['12:00', '13:00', '14:00'];
+const features = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const types = ['palace', 'flat', 'house', 'bungalow', 'hotel'];
+const photos = ['https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg', 'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'];
+const quantityOfAnnouncements = 10;
+
+export {checkins, checkouts, features, types, photos, quantityOfAnnouncements};
